modularize_v3: modularized CCSync related files
