+++ lib/app/modules/reports/views/burn_down_monthly_taskc.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:syncfusion_flutter_charts/charts.dart';
-import 'package:taskwarrior/api_service.dart';
 import 'package:taskwarrior/app/models/chart.dart';
 import 'package:taskwarrior/app/modules/reports/views/common_chart_indicator.dart';
 import 'package:taskwarrior/app/utils/app_settings/app_settings.dart';
@@ -10,6 +9,8 @@ import 'package:taskwarrior/app/utils/constants/taskwarrior_fonts.dart';
 import 'package:taskwarrior/app/utils/constants/utilites.dart';
 import 'package:taskwarrior/app/utils/themes/theme_extension.dart';
 import 'package:taskwarrior/app/utils/language/sentence_manager.dart';
+import 'package:taskwarrior/app/v3/db/task_database.dart';
+import 'package:taskwarrior/app/v3/models/task.dart';
 
 class BurnDownMonthlyTaskc extends StatelessWidget {
   BurnDownMonthlyTaskc({super.key});
@@ -33,22 +34,16 @@ class BurnDownMonthlyTaskc extends StatelessWidget {
           mainAxisSize: MainAxisSize.min,
           children: [
             Text(
-              '${SentenceManager(currentLanguage: AppSettings.selectedLanguage)
-                      .sentences
-                      .reportsMonthYear}: $monthYear',
+              '${SentenceManager(currentLanguage: AppSettings.selectedLanguage).sentences.reportsMonthYear}: $monthYear',
               style: const TextStyle(
                 fontWeight: TaskWarriorFonts.bold,
               ),
             ),
             Text(
-              '${SentenceManager(currentLanguage: AppSettings.selectedLanguage)
-                      .sentences
-                      .reportsPending}: $pendingCount',
+              '${SentenceManager(currentLanguage: AppSettings.selectedLanguage).sentences.reportsPending}: $pendingCount',
             ),
             Text(
-              '${SentenceManager(currentLanguage: AppSettings.selectedLanguage)
-                      .sentences
-                      .reportsCompleted}: $completedCount',
+              '${SentenceManager(currentLanguage: AppSettings.selectedLanguage).sentences.reportsCompleted}: $completedCount',
             ),
           ],
         ),
@@ -59,11 +54,11 @@ class BurnDownMonthlyTaskc extends StatelessWidget {
   Future<Map<String, Map<String, int>>> fetchMonthlyInfo() async {
     TaskDatabase taskDatabase = TaskDatabase();
     await taskDatabase.open();
-    List<Tasks> tasks = await taskDatabase.fetchTasksFromDatabase();
+    List<TaskForC> tasks = await taskDatabase.fetchTasksFromDatabase();
     return sortBurnDownMonthly(tasks);
   }
 
-  Map<String, Map<String, int>> sortBurnDownMonthly(List<Tasks> allData) {
+  Map<String, Map<String, int>> sortBurnDownMonthly(List<TaskForC> allData) {
     Map<String, Map<String, int>> monthlyInfo = {};
 
     allData.sort((a, b) => a.entry.compareTo(b.entry));
@@ -96,7 +91,8 @@ class BurnDownMonthlyTaskc extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final double height = MediaQuery.of(context).size.height;
-    TaskwarriorColorTheme tColors = Theme.of(context).extension<TaskwarriorColorTheme>()!;
+    TaskwarriorColorTheme tColors =
+        Theme.of(context).extension<TaskwarriorColorTheme>()!;
     return FutureBuilder<Map<String, Map<String, int>>>(
       future: fetchMonthlyInfo(),
       builder: (context, snapshot) {
@@ -106,10 +102,8 @@ class BurnDownMonthlyTaskc extends StatelessWidget {
 
         if (snapshot.hasError) {
           return Center(
-              child: Text('${SentenceManager(
-                          currentLanguage: AppSettings.selectedLanguage)
-                      .sentences
-                      .reportsError}: ${snapshot.error}'));
+              child: Text(
+                  '${SentenceManager(currentLanguage: AppSettings.selectedLanguage).sentences.reportsError}: ${snapshot.error}'));
         }
 
         Map<String, Map<String, int>> monthlyInfo = snapshot.data ?? {};
