+++ lib/app/v3/models/annotation.dart
@@ -1,0 +1,5 @@
+class Annotation {
+  final String? entry;
+  final String? description;
+  Annotation({this.entry, this.description});
+}
