+++ lib/app/v3/net/origin.dart
@@ -1,0 +1,1 @@
+String origin = 'http://localhost:8080';
